Changed type of 'range' columns to integers.
From now on all distances will be calculating in meters.

+++ app/models/classes/GPSTasks.scala
@@ -25,7 +25,7 @@ object GPSTasks extends Table[GPSTask]("GPSTASKS") {
   def pointId = column[Int]("pointId", O.NotNull)
   def lat   = column[Double]("lat", O.NotNull)
   def lon   = column[Double]("lon", O.NotNull)
-  def range = column[Double]("range", O.NotNull)
+  def range = column[Int]("range", O.NotNull)
   def * = gameId ~ taskId ~ pointId ~ lat ~ lon ~ range <> (GPSTask, GPSTask.unapply _)
   def pk = primaryKey("GPSTASKS_PK", (gameId, taskId, pointId))
   def game = foreignKey("GPSTASKS_GAMES_FK", gameId, Games)(_.id)
